requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
jieba>=0.42.1
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
numpy>=1.24.0
pandas>=2.0.0
xlsxwriter>=3.1.0
opencc-python-reimplemented>=0.1.7
chardet>=5.2.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# New feature dependencies
# Text similarity analysis
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Advanced visualizations
plotly>=5.15.0
networkx>=3.1
dash>=2.14.0

# GPU acceleration (optional)
accelerate>=0.20.0

# Task queue system
celery>=5.3.0
redis>=4.6.0

# Extended file format support
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
requests>=2.31.0
pdfplumber>=0.9.0 